plagiarism_checker_proj-clq/requirements.txt
sentence-transformers
faiss-cpu
numpy
streamlit